requirements.txt
transformers 
transformers[sentencepiece]
datasets 
sacrebleu
rouge_score 
py7zr
pandas 
nltk 
tqdm
PyYAML
matplotlib
torch 
notebook
boto3
mypy-boto3-s3 
python-box 
ensure 
fastapi 
uvicorn
jinja2 
-e .


# https://github.com/entbappy/Branching-tutorial/raw/master/samsumdata.zip